+++ src/CloudMigrator.Cli/Program.cs
@@ -11,6 +11,9 @@ Console.CancelKeyPress += (_, e) =>
 var rootCmd = new RootCommand("CloudMigrator - OneDrive から SharePoint へのファイル移行ツール");
 rootCmd.Add(TransferCommand.Build());
 rootCmd.Add(RebuildSkipListCommand.Build());
+rootCmd.Add(WatchdogCommand.Build());
+rootCmd.Add(QualityMetricsCommand.Build());
+rootCmd.Add(SecurityScanCommand.Build());
 
 return await rootCmd.Parse(args).InvokeAsync(new InvocationConfiguration(), cts.Token);